# docs/openapi.yaml
openapi: 3.1.0
info:
  title: CityConnect API
  version: "1.0.0"
  description: |
    Municipal CRM and service-request API.

    Three kinds of caller reach this API, and they authenticate differently:

    - **Staff** hold an HTTP-only session cookie, established by signing in
      through C2 (`/auth/login` → C2 → `/auth/callback`). There is no local
      account path.
    - **Connected systems** present a `cc_pat_…` personal access token as a
      bearer credential. A token can only ever narrow its owner's authority.
    - **C2 itself** calls two endpoints server-to-server: the Service Card
      callout (`/citizens/{sub}/status`, authenticated by a short-lived RS256
      assertion) and back-channel logout (a signed `logout_token`).

    Errors are `application/problem+json` and carry a stable `code` field.
    Switch on that rather than on the prose.
  contact:
    name: CityConnect
    url: https://github.com/jjamieson1/CityConnect
servers:
  - url: https://city.example.gov/cityconnect/api
    description: Production
  - url: http://localhost:4021/api
    description: Local development
...
security:
  - sessionCookie: []
  - bearerToken: []
paths:
...
  /auth/login:
    get:
      tags: [Auth]
      summary: Begin sign-in
      security: []
      description: |
        Redirects the browser to C2's authorization endpoint with PKCE.

        Deliberately sends neither `prompt=login` nor `max_age` — either one
        makes C2 re-prompt for credentials on every visit even with an active
        session, destroying SSO.
      parameters:
        - { name: returnTo, in: query, schema: { type: string }, description: "Path to return to; must be relative" }
        - { name: silent, in: query, schema: { type: boolean }, description: "Send prompt=none for a background session check" }
      responses:
        "302": { description: Redirect to C2 }
        "200":
          description: The authorize URL, when the caller asked for JSON
          content:
            application/json:
              schema:
                type: object
                properties: { authorizeUrl: { type: string } }
  /auth/callback:
    get:
      tags: [Auth]
      summary: Complete sign-in
      security: []
      description: Validates state, exchanges the code, verifies the id_token, and sets the session cookie.
      parameters:
        - { name: code, in: query, schema: { type: string } }
        - { name: state, in: query, schema: { type: string } }
        - { name: error, in: query, schema: { type: string } }
      responses:
        "302": { description: Redirect into the console, or back to sign-in with a reason }
...
  /auth/me:
    get:
      tags: [Auth]
      summary: Current identity and capabilities
      description: |
        Returns the caller plus the permission list the console renders from,
        so a permission change on the server reaches the UI without a front-end
        release.
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Me" }
        "401": { $ref: "#/components/responses/Unauthorized" }

  # ---------------------------------------------------------------------------
  /citizens/{sub}/status:
    get:
      tags: [C2]
      summary: Service Card callout
      security:
        - calloutAssertion: []
      description: |
        C2 calls this server-to-server when a citizen opens their Service Card,
        and renders the returned JSON.

        Constraints worth respecting: C2 allows about five seconds, does not
        cache, calls on every render plus a refresh timer, and falls back to
        static card content on any error — so a clean 4xx is better than a
        malformed 200. An unknown subject returns `200 {}`, never a guess.
      parameters:
        - name: sub
          in: path
          required: true
          schema: { type: string }
          description: The citizen's opaque C2 subject identifier
      responses:
        "200":
          description: The status bundle
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StatusBundle" }
        "401": { description: The assertion failed verification }
        "503": { description: Temporarily unavailable; C2 shows the static card }
  /c2/backchannel-logout:
    post:
      tags: [C2]
      summary: Back-channel logout
      security: []
      description: |
        C2 posts a signed `logout_token` when a user explicitly signs out.

        The token is subject-based and carries no `sid`, so a single session
        cannot be targeted — every session that subject holds is ended.
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [logout_token]
              properties: { logout_token: { type: string } }
      responses:
        "200": { description: Sessions terminated }
        "400": { description: The token failed validation }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: cc_session
      description: Set by the C2 sign-in callback. HTTP-only, SameSite=Lax.
    bearerToken:
      type: http
      scheme: bearer
      description: A `cc_pat_…` personal access token, for connected systems.
    calloutAssertion:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        C2's short-lived RS256 Service Card assertion, verified against the same
        JWKS as id_tokens. `aud` must equal our `client_id`; `exp` is about 60
        seconds after `iat`.
...
  responses:
...
    Unauthorized:
      description: No valid session or token
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
  schemas:
    Problem:
      type: object
      description: RFC 7807. Switch on `code`, which is stable; `detail` is prose.
      properties:
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        code:
          type: string
          enum:
            [not_found, unauthenticated, no_access, suspended, forbidden,
             stale_version, bad_transition, duplicate, conflict, suppressed,
             invalid_input, invalid_json, rate_limited, internal]
        instance: { type: string }
        requestId: { type: string }
...
    StatusBundle:
      type: object
      description: |
        What C2 renders on a Service Card. Every field is optional; C2 falls
        back to the admin-configured content for anything omitted.
      properties:
        title: { type: string }
        description: { type: string }
        CTA:
          type: string
          description: |
            Note the capitalisation — that is C2's contract, not a typo.
            Ignored when `tasks` is present.
        contact:
          type: object
          properties:
            email: { type: string }
            phone: { type: string }
            address1: { type: string }
            city: { type: string }
            postalCode: { type: string }
        tasks:
          type: array
          items:
            type: object
            properties:
              name: { type: string }
              description: { type: string }
              url: { type: string, description: "Absolute; opened in a new tab." }
    Me:
      type: object
      properties:
        user: { type: object }
        department: { type: object }
        queues: { type: array, items: { type: object } }
        permissions:
          type: array
          items: { type: string }
          description: The console renders from this rather than duplicating the role table.
        crossDepartment: { type: boolean }